lesson 1 made C

+++ Занятие 1/Практические задания/!ProgramTasks/C/Program.cs
@@ -5,7 +5,7 @@
         static void Main(string[] args)
         {
             double pi = Math.PI;
-            int tenThousand = (int)10000L;
+            long tenThousand = 10000L;
             float tenThousandPi = (float)(pi * tenThousand);
             int roundedTenThousandPi = (int)Math.Round(tenThousandPi);
             int integerPartOfTenThousandPi = (int)tenThousandPi;
